fix(approvals): bypass approval for background job control
